Revert "Using persistent Envoy and ext_proc server in ext_proc fuzzing test. (#26499)"
This reverts commit 4a2e237ecbdc66ea8e18ad92c1d08dbe23dd99bc.
