+++ test/extensions/filters/http/ext_proc/BUILD
@@ -211,36 +211,25 @@ envoy_proto_library(
     srcs = ["ext_proc_grpc_fuzz.proto"],
 )
 
-EXT_PROC_GRPC_FUZZ_TEST_DEPS = [
-    ":ext_proc_grpc_fuzz_lib",
-    ":ext_proc_grpc_fuzz_proto_cc_proto",
-    ":test_processor_lib",
-    "//source/common/network:address_lib",
-    "//source/extensions/filters/http/ext_proc:config",
-    "//test/common/http:common_lib",
-    "//test/fuzz:utility_lib",
-    "//test/integration:http_integration_lib",
-    "//test/test_common:utility_lib",
-    "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-    "@envoy_api//envoy/extensions/filters/http/ext_proc/v3:pkg_cc_proto",
-    "@envoy_api//envoy/service/ext_proc/v3:pkg_cc_proto",
-    "@envoy_api//envoy/type/v3:pkg_cc_proto",
-]
-
 envoy_cc_fuzz_test(
     name = "ext_proc_grpc_fuzz_test",
     srcs = ["ext_proc_grpc_fuzz.cc"],
-    hdrs = ["ext_proc_grpc_fuzz.h"],
     corpus = "ext_proc_grpc_corpus",
-    deps = EXT_PROC_GRPC_FUZZ_TEST_DEPS,
-)
-
-envoy_cc_fuzz_test(
-    name = "ext_proc_grpc_fuzz_test_persistent",
-    srcs = ["ext_proc_grpc_fuzz_persistent.cc"],
-    hdrs = ["ext_proc_grpc_fuzz.h"],
-    corpus = "ext_proc_grpc_corpus",
-    deps = EXT_PROC_GRPC_FUZZ_TEST_DEPS,
+    deps = [
+        ":ext_proc_grpc_fuzz_lib",
+        ":ext_proc_grpc_fuzz_proto_cc_proto",
+        ":test_processor_lib",
+        "//source/common/network:address_lib",
+        "//source/extensions/filters/http/ext_proc:config",
+        "//test/common/http:common_lib",
+        "//test/fuzz:utility_lib",
+        "//test/integration:http_integration_lib",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/filters/http/ext_proc/v3:pkg_cc_proto",
+        "@envoy_api//envoy/service/ext_proc/v3:pkg_cc_proto",
+        "@envoy_api//envoy/type/v3:pkg_cc_proto",
+    ],
 )
 
 envoy_extension_cc_test(
