+++ test/extensions/filters/http/ext_proc/ext_proc_grpc_fuzz_helper.h
@@ -106,9 +106,13 @@ public:
   void randomizeImmediateResponse(ImmediateResponse* msg, const ProcessingRequest* req);
   void randomizeOverrideResponse(ProcessingMode* msg);
   void randomizeResponse(ProcessingResponse* resp, const ProcessingRequest* req);
-  grpc::Status generateResponse(ProcessingRequest& req, ProcessingResponse& resp,
-                                bool& immediate_close_grpc);
+
   FuzzedDataProvider* provider_;
+
+  // Protects immediate_resp_sent_
+  Thread::MutexBasicLockable immediate_resp_lock_;
+  // Flags if an immediate response was generated and sent
+  bool immediate_resp_sent_;
 };
 
 } // namespace ExternalProcessing
