--- test/extensions/filters/http/ext_proc/ext_proc_grpc_fuzz_persistent.cc
@@ -1,15 +1,0 @@
-#include "test/extensions/filters/http/ext_proc/ext_proc_grpc_fuzz.h"
-
-namespace Envoy {
-namespace Extensions {
-namespace HttpFilters {
-namespace ExternalProcessing {
-
-DEFINE_PROTO_FUZZER(const test::extensions::filters::http::ext_proc::ExtProcGrpcTestCase& input) {
-  fuzzExtProcRun(input, true);
-}
-
-} // namespace ExternalProcessing
-} // namespace HttpFilters
-} // namespace Extensions
-} // namespace Envoy
